uboot: Copy additional DT property for pi400 PHY
THe kernel would not be able to enable the ethernet PHY device because of a discrepancy between the actual PHY address as provided by the firmware and the address specified in the kernel DTB for the Pi 400.

This additional patch builds on the previous one, carrying over the relevant property.

Signed-off-by: Antoine Mazeas <antoine@karthanis.net>